fix: error page redirect, default password and image, seeder for admin remove : role superuser

+++ database/migrations/2025_02_03_080544_create_bills_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('student_id');
             $table->unsignedBigInteger('payment_type_id');
-            $table->integer('amount');
+            $table->decimal('amount', 12, 2);
             $table->enum('status', ['pending', 'paid'])->default('pending');
             $table->timestamps();
